requirements.txt
streamlit
python-dotenv
langchain-text-splitters
langchain-community
langgraph
langchain
langchain-huggingface
langchain-core
langchain-chroma
langchain-pinecone
bs4
chromadb
pinecone-client>=3.0.0
groq
requests
pandas
numpy
ipykernel
tqdm
matplotlib
seaborn
jupyter